qubitcode_backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.4
Werkzeug==2.3.7
gunicorn==21.2.0